fix: remove duplicate code
remove duplicate function which had same code and update tests
